feat: main authorization dispaly add & fcm edit
