+++ src/pages/Main/style.scss
@@ -1,40 +1,218 @@
 @import "../../common/styles/scss/responsive.scss";
+.spinnerContainer {
+  @apply flex flex-col place-items-center my-auto;
+  //display: flex;
+  //flex-direction: column;
+  //align-items: center;
+}
+
+.spinner {
+  width: 56px;
+  height: 56px;
+  display: grid;
+  border: 4px solid #0000;
+  border-radius: 50%;
+  border-right-color: #299fff;
+  animation: tri-spinner 1s infinite linear;
+}
+
+.spinner::before,
+.spinner::after {
+  content: "";
+  grid-area: 1/1;
+  margin: 2px;
+  border: inherit;
+  border-radius: 50%;
+  animation: tri-spinner 2s infinite;
+}
+
+.spinner::after {
+  margin: 8px;
+  animation-duration: 3s;
+}
+
+@keyframes tri-spinner {
+  100% {
+    transform: rotate(1turn);
+  }
+}
+
 .loader {
-  //   position: relative;
-  @include responsive-position(200px, 50px);
+  color: #4a4a4a;
+  font-family: "Poppins", sans-serif;
+  font-weight: 500;
+  font-size: 25px;
+  -webkit-box-sizing: content-box;
+  box-sizing: content-box;
+  height: 40px;
+  padding: 10px 10px;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  border-radius: 8px;
+  font-family: "Hana2M";
+}
+
+.words {
   overflow: hidden;
-  border-right: 3px solid;
-  width: 0px;
-  animation:
-    typewriter 2s steps(10) infinite alternate,
-    blink 0.5s steps(10) infinite;
 }
 
-.loader-text {
-  //   font-size: 40px;
-  @include responsive-font-size(60px);
-  font-weight: 700;
-  background: linear-gradient(to right, #159957, #155799);
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: transparent;
+.word {
+  display: block;
+  height: 100%;
+  padding-left: 6px;
+  // color: #299fff;
+  color: #44bd91;
+  animation: cycle-words 5s infinite;
+  font-family: "Hana2M";
 }
 
-@keyframes typewriter {
-  0% {
-    width: 0px;
+@keyframes cycle-words {
+  10% {
+    -webkit-transform: translateY(-105%);
+    transform: translateY(-105%);
+  }
+
+  25% {
+    -webkit-transform: translateY(-100%);
+    transform: translateY(-100%);
+  }
+
+  35% {
+    -webkit-transform: translateY(-205%);
+    transform: translateY(-205%);
+  }
+
+  50% {
+    -webkit-transform: translateY(-200%);
+    transform: translateY(-200%);
+  }
+
+  60% {
+    -webkit-transform: translateY(-305%);
+    transform: translateY(-305%);
+  }
+
+  75% {
+    -webkit-transform: translateY(-300%);
+    transform: translateY(-300%);
+  }
+
+  85% {
+    -webkit-transform: translateY(-405%);
+    transform: translateY(-405%);
   }
 
   100% {
-    width: 240px;
+    -webkit-transform: translateY(-400%);
+    transform: translateY(-400%);
   }
 }
 
-@keyframes blink {
+// loading - 2
+.typing-indicator {
+  @include responsive-position(300px, 140px);
+  @include responsive-hw(30px, 150px);
+  // width: 60px;
+  // height: 30px;
+  // position: relative;
+  z-index: 4;
+}
+
+.typing-circle {
+  // width: 8px;
+  // height: 8px;
+  @include responsive-hw(15px, 15px);
+  position: absolute;
+  border-radius: 50%;
+  // background-color: #000;
+  background-color: #44bd91;
+  left: 15%;
+  transform-origin: 50%;
+  animation: typing-circle7124 0.5s alternate infinite ease;
+}
+
+.mainTitle {
+  @apply grid text-center mt-96;
+  @include responsive-font-size(20px);
+  font-family: "NanumSquareRoundR";
+  line-height: 1.5;
+}
+
+.mainSub {
+  @apply grid text-center mt-80;
+  @include responsive-font-size(25px);
+  color: #44bd91;
+  font-family: "Hana2M";
+}
+
+@keyframes typing-circle7124 {
   0% {
-    border-right-color: rgba(255, 255, 255, 0.75);
+    top: 20px;
+    height: 5px;
+    border-radius: 50px 50px 25px 25px;
+    transform: scaleX(1.7);
+  }
+
+  40% {
+    height: 8px;
+    border-radius: 50%;
+    transform: scaleX(1);
   }
 
   100% {
-    border-right-color: transparent;
+    top: 0%;
   }
+}
+
+.typing-circle:nth-child(2) {
+  left: 45%;
+  animation-delay: 0.2s;
+}
+
+.typing-circle:nth-child(3) {
+  left: auto;
+  right: 15%;
+  animation-delay: 0.3s;
+}
+
+.typing-shadow {
+  width: 5px;
+  height: 4px;
+  border-radius: 50%;
+  background-color: rgba(0, 0, 0, 0.2);
+  position: absolute;
+  top: 30px;
+  transform-origin: 50%;
+  z-index: 3;
+  left: 15%;
+  filter: blur(1px);
+  animation: typing-shadow046 0.5s alternate infinite ease;
+}
+
+@keyframes typing-shadow046 {
+  0% {
+    transform: scaleX(1.5);
+  }
+
+  40% {
+    transform: scaleX(1);
+    opacity: 0.7;
+  }
+
+  100% {
+    transform: scaleX(0.2);
+    opacity: 0.4;
+  }
+}
+
+.typing-shadow:nth-child(4) {
+  left: 45%;
+  animation-delay: 0.2s;
+}
+
+.typing-shadow:nth-child(5) {
+  left: auto;
+  right: 15%;
+  animation-delay: 0.3s;
 }
